+++ examples/9_particle_write_serial.py
@@ -6,8 +6,8 @@ Copyright 2019-2020 openPMD contributors
 Authors: Axel Huebl
 License: LGPLv3+
 """
-from openpmd_api import Series, Access_Type, Dataset, Mesh_Record_Component, \
-    Unit_Dimension
+from openpmd_api import Series, Access_Type, Dataset, Flush_Type, \
+    Mesh_Record_Component, Unit_Dimension
 import numpy as np
 
 
@@ -20,6 +20,9 @@ if __name__ == "__main__":
         "../samples/7_particle_write_serial_py.h5",
         Access_Type.create
     )
+
+    # FIXME GitHub issue #490
+    f.set_flush(Flush_Type.defer)
 
     # all required openPMD attributes will be set to reasonable default values
     # (all ones, all zeros, empty strings,...)